requirements.txt
python-pdfbox
beautifulsoup4
fake-useragent
selenium
pypdf
huggingface_hub
acryl-datahub[datahub-rest]
datahub
torch
transformers
python-dotenv
langchain
tiktoken
webdriver-manager
mlflow==2.4.1
greenplum-python
torchvision
accelerate

matplotlib
pmdarima
prophet
psycopg2-binary
sqlalchemy
python-dotenv
pandas
wheel
scikit-learn
nbconvert
seaborn
traitlets
scipy
plotly
watchdog
mpld3
streamlit==1.18.0
Pillow
streamlit-autorefresh
jupyterflow
lz4
pyyaml==5.3.1
prometheus-client
protobuf # protobuf==3.20.0 # protobuf==3.19.5
fastapi
uvicorn[standard]
gunicorn
boto3
tensorflow
rsocket
python-snappy
filelock==3.9.0
pyarrow<7.0.0
hickle
python-multipart
evidently
#pysqlite3-binary
kfp==2.5.0
python-multipart
pyservicebinding
download_and_extract
altair==4.0
sentencepiece
urllib3<2
GitPython
pydantic==2.6.0
#onnxruntime-genai
https://files.pythonhosted.org/packages/97/44/b9f0190001c1a3afe117c19cd33bfae4512ba5f847e0fe340ca1d3cd54ff/onnxruntime_genai-0.1.0-cp310-cp310-manylinux_2_27_x86_64.manylinux_2_28_x86_64.whl
#--extra-index-url=https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-genai/pypi/simple/
#--extra-index=url=https://pypi.org/project/onnxruntime-genai/
#--pre
.